# app/src/api/generated/model/foodEntryEntity.ts
/**
 * Generated by orval v6.5.3 🍺
 * Do not edit manually.
 * portfolio Calories Tracker App
 * The portfolio Calories Tracker App API description
 * OpenAPI spec version: 1.0
 */

// eslint-disable-next-line @typescript-eslint/no-empty-interface
export interface FoodEntryEntity {}