# app/src/api/generated/model/foodEntrySerializer.ts
/**
 * Generated by orval v6.5.3 🍺
 * Do not edit manually.
 * portfolio Calories Tracker App
 * The portfolio Calories Tracker App API description
 * OpenAPI spec version: 1.0
 */

export interface FoodEntrySerializer {
  id: number;
  food_name: string;
  taken_at: string;
  price: string;
  calories_count: string;
}