# app/src/api/generated/model/getManyDailyCalorieSerializer.ts
/**
 * Generated by orval v6.5.3 🍺
 * Do not edit manually.
 * portfolio Calories Tracker App
 * The portfolio Calories Tracker App API description
 * OpenAPI spec version: 1.0
 */
import type { DailyCalorieSerializer } from './dailyCalorieSerializer';

export interface GetManyDailyCalorieSerializer {
  count: number;
  total: number;
  page: number;
  pageCount: number;
  data: DailyCalorieSerializer;
}