# app/src/api/generated/model/monthlySpendingSerializer.ts
/**
 * Generated by orval v6.5.3 🍺
 * Do not edit manually.
 * portfolio Calories Tracker App
 * The portfolio Calories Tracker App API description
 * OpenAPI spec version: 1.0
 */

export interface MonthlySpendingSerializer {
  user_id: number;
  month: string;
  spending: number;
}