# app/src/api/generated/model/userProfileSerializer.ts
/**
 * Generated by orval v6.5.3 🍺
 * Do not edit manually.
 * portfolio Calories Tracker App
 * The portfolio Calories Tracker App API description
 * OpenAPI spec version: 1.0
 */

export interface UserProfileSerializer {
  id: number;
  first_name: string;
  last_name: string;
  email: string;
  user_calories_limit: number;
  user_monthly_spend_limit: number;
}