# packages/backend/swagger-spec.json
{"openapi":"3.0.0","paths":{"/login":{"post":{"operationId":"login","summary":"","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"Return user access token.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccessTokenSerializer"}}}}}}},"/register":{"post":{"operationId":"register","summary":"","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"200":{"description":"Return user access token.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccessTokenSerializer"}}}}}}},"/me":{"get":{"operationId":"me","summary":"","parameters":[],"responses":{"200":{"description":"Get user profile.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserProfileSerializer"}}}}},"security":[{"bearer":[]}]}},"/food-entries":{"post":{"operationId":"createFoodEntry","summary":"","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateFoodEntryDto"}}}},"responses":{"201":{"description":"Get create one base response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FoodEntrySerializer"}}}}},"security":[{"bearer":[]}]},"get":{"operationId":"getFoodEntries","summary":"","parameters":[{"name":"fields","description":"Selects resource fields. <a href=\"https://github.com/nestjsx/crud/wiki/Requests#select\" target=\"_blank\">Docs</a>","required":false,"in":"query","schema":{"type":"array","items":{"type":"string"}},"style":"form","explode":false},{"name":"s","description":"Adds search condition. <a href=\"https://github.com/nestjsx/crud/wiki/Requests#search\" target=\"_blank\">Docs</a>","required":false,"in":"query","schema":{"type":"string"}},{"name":"filter","description":"Adds filter condition. <a href=\"https://github.com/nestjsx/crud/wiki/Requests#filter\" target=\"_blank\">Docs</a>","required":false,"in":"query","schema":{"type":"array","items":{"type":"string"}},"style":"form","explode":true},{"name":"or","description":"Adds OR condition. <a href=\"https://github.com/nestjsx/crud/wiki/Requests#or\" target=\"_blank\">Docs</a>","required":false,"in":"query","schema":{"type":"array","items":{"type":"string"}},"style":"form","explode":true},{"name":"sort","description":"Adds sort by field. <a href=\"https://github.com/nestjsx/crud/wiki/Requests#sort\" target=\"_blank\">Docs</a>","required":false,"in":"query","schema":{"type":"array","items":{"type":"string"}},"style":"form","explode":true},{"name":"join","description":"Adds relational resources. <a href=\"https://github.com/nestjsx/crud/wiki/Requests#join\" target=\"_blank\">Docs</a>","required":false,"in":"query","schema":{"type":"array","items":{"type":"string"}},"style":"form","explode":true},{"name":"limit","description":"Limit amount of resources. <a href=\"https://github.com/nestjsx/crud/wiki/Requests#limit\" target=\"_blank\">Docs</a>","required":false,"in":"query","schema":{"type":"integer"}},{"name":"offset","description":"Offset amount of resources. <a href=\"https://github.com/nestjsx/crud/wiki/Requests#offset\" target=\"_blank\">Docs</a>","required":false,"in":"query","schema":{"type":"integer"}},{"name":"page","description":"Page portion of resources. <a href=\"https://github.com/nestjsx/crud/wiki/Requests#page\" target=\"_blank\">Docs</a>","required":false,"in":"query","schema":{"type":"integer"}},{"name":"cache","description":"Reset cache (if was enabled). <a href=\"https://github.com/nestjsx/crud/wiki/Requests#cache\" target=\"_blank\">Docs</a>","required":false,"in":"query","schema":{"type":"integer","minimum":0,"maximum":1}}],"responses":{"200":{"description":"Get paginated response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetManyFoodEntriesSerializer"}}}}},"security":[{"bearer":[]}]}},"/food-entries/monthly-spending":{"get":{"operationId":"getMonthlySpending","summary":"","parameters":[{"name":"filter","required":false,"in":"query","schema":{"type":"array","items":{"type":"string"}}},{"name":"sort","required":false,"in":"query","schema":{"type":"array","items":{"type":"string"}}},{"name":"page","required":false,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":"Get user profile.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetManyMonthlySpendingSerializer"}}}}},"security":[{"bearer":[]}]}},"/food-entries/daily-calories":{"get":{"operationId":"getDailyCalories","summary":"","parameters":[{"name":"filter","required":false,"in":"query","schema":{"type":"array","items":{"type":"string"}}},{"name":"sort","required":false,"in":"query","schema":{"type":"array","items":{"type":"string"}}},{"name":"page","required":false,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":"Get user profile.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetManyDailyCalorieSerializer"}}}}},"security":[{"bearer":[]}]}}},"info":{"title":"portfolio Calories Tracker App","description":"The portfolio Calories Tracker App API description","version":"1.0","contact":{}},"tags":[{"name":"calories-tracker","description":""}],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"LoginDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"AccessTokenSerializer":{"type":"object","properties":{"access_token":{"type":"string"}},"required":["access_token"]},"RegisterDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"first_name":{"type":"string"},"last_name":{"type":"string"}},"required":["email","password","first_name","last_name"]},"UserProfileSerializer":{"type":"object","properties":{"id":{"type":"number"},"first_name":{"type":"string"},"last_name":{"type":"string"},"email":{"type":"string"},"user_calories_limit":{"type":"number"},"user_monthly_spend_limit":{"type":"number"}},"required":["id","first_name","last_name","email","user_calories_limit","user_monthly_spend_limit"]},"CreateFoodEntryDto":{"type":"object","properties":{"food_name":{"type":"string"},"taken_at":{"type":"string"},"calories_count":{"type":"number"},"price":{"type":"number"}},"required":["food_name","taken_at","calories_count","price"]},"GetManyFoodEntriesSerializer":{"type":"object","properties":{"count":{"type":"number"},"total":{"type":"number"},"page":{"type":"number"},"pageCount":{"type":"number"},"data":{"$ref":"#/components/schemas/FoodEntrySerializer"}},"required":["count","total","page","pageCount","data"]},"FoodEntryEntity":{"type":"object","properties":{}},"FoodEntrySerializer":{"type":"object","properties":{"id":{"type":"number"},"food_name":{"type":"string"},"taken_at":{"type":"string"},"price":{"type":"string"},"calories_count":{"type":"string"}},"required":["id","food_name","taken_at","price","calories_count"]},"MonthlySpendingSerializer":{"type":"object","properties":{"user_id":{"type":"number"},"month":{"type":"string"},"spending":{"type":"number"}},"required":["user_id","month","spending"]},"GetManyMonthlySpendingSerializer":{"type":"object","properties":{"count":{"type":"number"},"total":{"type":"number"},"page":{"type":"number"},"pageCount":{"type":"number"},"data":{"$ref":"#/components/schemas/MonthlySpendingSerializer"}},"required":["count","total","page","pageCount","data"]},"DailyCalorieSerializer":{"type":"object","properties":{"user_id":{"type":"number"},"date":{"type":"string"},"calories_count":{"type":"number"}},"required":["user_id","date","calories_count"]},"GetManyDailyCalorieSerializer":{"type":"object","properties":{"count":{"type":"number"},"total":{"type":"number"},"page":{"type":"number"},"pageCount":{"type":"number"},"data":{"$ref":"#/components/schemas/DailyCalorieSerializer"}},"required":["count","total","page","pageCount","data"]}}}}